requirements.txt
django==1.11.*
psycopg2-binary==2.8.*
environs==9.3.3
python-dotenv==0.19.0
